+++ gql/lib/execution.dart
@@ -94,14 +94,19 @@ class Response {
   /// Follows the shape of requested document.
   final Map<String, dynamic> data;
 
+  /// A [Context] to be returned along with a [Response]
+  final Context context;
+
   const Response({
     this.errors,
     this.data,
-  });
+    this.context = const Context(),
+  }) : assert(context != null);
 
   List<Object> _getChildren() => [
         errors,
         data,
+        context,
       ];
 
   @override
@@ -172,7 +177,8 @@ class Request {
   const Request({
     @required this.operation,
     this.context = const Context(),
-  }) : assert(operation != null);
+  })  : assert(operation != null),
+        assert(context != null);
 
   List<Object> _getChildren() => [
         operation,